+++ controllers/om/backup/backup.go
@@ -64,7 +64,7 @@ type HostClusterReader interface {
 
 // StopBackupIfEnabled tries to find backup configuration for specified resource (can be Replica Set or Sharded Cluster -
 // Ops Manager doesn't backup Standalones) and disable it.
-func StopBackupIfEnabled(readUpdater ConfigReadUpdater, hostClusterReader HostClusterReader, name string, resourceType MongoDbResourceType, log *zap.SugaredLogger) error {
+func StopBackupIfEnabled(readUpdater ConfigHostReadUpdater, hostClusterReader HostClusterReader, name string, resourceType MongoDbResourceType, log *zap.SugaredLogger) error {
 	response, err := readUpdater.ReadBackupConfigs()
 	if err != nil {
 		// If the operator can't read BackupConfigs, it might indicate that the Pods were removed before establishing
@@ -110,7 +110,7 @@ func StopBackupIfEnabled(readUpdater ConfigReadUpdater, hostClusterReader HostCl
 	return nil
 }
 
-func disableBackup(readUpdater ConfigReadUpdater, backupConfig *Config, log *zap.SugaredLogger) error {
+func disableBackup(readUpdater ConfigHostReadUpdater, backupConfig *Config, log *zap.SugaredLogger) error {
 	if backupConfig.Status == Started {
 		err := readUpdater.UpdateBackupStatus(backupConfig.ClusterId, Stopped)
 		if err != nil {