+++ controllers/operator/common_controller.go
@@ -417,7 +417,7 @@ func (r *ReconcileCommonController) ensureX509AgentCertsForMongoDBResource(mdb *
 
 // ensureBackupConfigurationAndUpdateStatus configures backup in Ops Manager based on the MongoDB resources spec
 func (r *ReconcileCommonController) ensureBackupConfigurationAndUpdateStatus(conn om.Connection, mdb *mdbv1.MongoDB, log *zap.SugaredLogger) workflow.Status {
-	statusOpt, opts := backup.EnsureBackupConfigurationInOpsManager(mdb.Spec.Backup, conn.GroupID(), conn, log)
+	statusOpt, opts := backup.EnsureBackupConfigurationInOpsManager(*mdb, conn.GroupID(), conn, log)
 	if len(opts) > 0 {
 		if _, err := r.updateStatus(mdb, statusOpt, log, opts...); err != nil {
 			return workflow.Failed(err.Error())